Got mouse making logical decisions around maze
Program does not work - Mouse needs to trace back through path. In order to do this a dynamic array list containing the vector positions of its visited path must be created. The algorithm that directs the mouse must then be completed to ensure the mouse will keep going back through the maze until it reaches a new tile.